+++ src/devicemr.js
@@ -118,7 +118,7 @@ class MerakiDevice extends EventEmitter {
                 networkId: this.networkId,
                 deviceData: this.deviceData,
                 hideUnconfiguredSsid: this.hideUnconfiguredSsids,
-                debugLog: this.enableDebugMode
+                enableDebugMode: this.enableDebugMode
             });
 
             this.merakiMr.on('deviceInfo', (ssidsCount) => {
@@ -151,11 +151,11 @@ class MerakiDevice extends EventEmitter {
                     };
                 }
             })
-                .on('success', (message) => {
-                    this.emit('success', message);
+                .on('success', (success) => {
+                    this.emit('success', success);
                 })
-                .on('message', (message) => {
-                    this.emit('message', message);
+                .on('info', (info) => {
+                    this.emit('message', info);
                 })
                 .on('debug', (debug) => {
                     this.emit('debug', debug);
@@ -165,10 +165,13 @@ class MerakiDevice extends EventEmitter {
                 })
                 .on('error', (error) => {
                     this.emit('error', error);
-                });
+                })
 
             //connect
-            await this.merakiMr.connect();
+            const connect = await this.merakiMr.connect();
+            if (!connect) {
+                return false;
+            };
 
             //prepare accessory
             if (this.startPrepareAccessory) {