+++ src/devicems.js
@@ -119,7 +119,7 @@ class MerakiDevice extends EventEmitter {
                 host: this.host,
                 apiKey: this.apiKey,
                 deviceData: this.deviceData,
-                debugLog: this.enableDebugMode
+                enableDebugMode: this.enableDebugMode
             });
 
             this.merakiMs.on('deviceInfo', (portsCount) => {
@@ -152,11 +152,11 @@ class MerakiDevice extends EventEmitter {
                     };
                 };
             })
-                .on('success', (message) => {
-                    this.emit('success', message);
+                .on('success', (success) => {
+                    this.emit('success', success);
                 })
-                .on('message', (message) => {
-                    this.emit('message', message);
+                .on('info', (info) => {
+                    this.emit('message', info);
                 })
                 .on('debug', (debug) => {
                     this.emit('debug', debug);
@@ -166,10 +166,13 @@ class MerakiDevice extends EventEmitter {
                 })
                 .on('error', (error) => {
                     this.emit('error', error);
-                });
+                })
 
             //connect
-            await this.merakiMs.connect();
+            const connect = await this.merakiMs.connect();
+            if (!connect) {
+                return false;
+            };
 
             //prepare accessory
             if (this.startPrepareAccessory) {